+++ pappl/device-usb.c
@@ -55,7 +55,7 @@ static void		pappl_usb_close(pappl_device_t *device);
 static bool		pappl_usb_find(pappl_device_cb_t cb, void *data, _pappl_usb_dev_t *device, pappl_deverror_cb_t err_cb, void *err_data);
 static char		*pappl_usb_getid(pappl_device_t *device, char *buffer, size_t bufsize);
 static bool		pappl_usb_list(pappl_devtype_t types, pappl_device_cb_t cb, void *data, pappl_deverror_cb_t err_cb, void *err_data);
-static bool		pappl_usb_open(pappl_device_t *device, const char *device_uri, const char *name);
+static bool		pappl_usb_open(pappl_device_t *device, const char *device_uri, pappl_job_t *job);
 static bool		pappl_usb_open_cb(const char *device_info, const char *device_uri, const char *device_id, void *data);
 static ssize_t		pappl_usb_read(pappl_device_t *device, void *buffer, size_t bytes);
 static pappl_preason_t	pappl_usb_status(pappl_device_t *device);
@@ -555,12 +555,12 @@ static bool				// `true` on success, `false` on error
 pappl_usb_open(
     pappl_device_t *device,		// I - Device
     const char     *device_uri,		// I - Device URI
-    const char     *job_name)		// I - Job name (unused)
+    pappl_job_t    *job)		// I - Job (unused)
 {
   _pappl_usb_dev_t	*usb;		// USB device
 
 
-  (void)job_name;
+  (void)job;
 
   if ((usb = (_pappl_usb_dev_t *)calloc(1, sizeof(_pappl_usb_dev_t))) == NULL)
   {