+++ CHANGES.md
@@ -23,4 +23,6 @@ Changes in v2.0b1
 - Fixed potential deadlock issue (Issue #297)
 - Fixed loading of previous state (Issue #298)
 - Fixed "printer-id" value for new printers (Issue #301)
+- Fixed Set-Printer-Attributes for "output-bin-default" and "sides-default"
+  (Issue #305) 
 - Fixed default "copies" value with `papplJobCreateWithFile`.